requirements.txt
requests
beautifulsoup4
Flask
Flask-Cors
python-dotenv
google-generativeai
google-api-core